requirements.txt
astroid==1.4.8
Bottleneck==1.2.0
deap==1.0.2
decorator==4.0.6
isort==4.2.5
lazy-object-proxy==1.2.2
llvmlite==0.14.0
mccabe==0.5.2
mlxtend==0.5.0
nose==1.3.7
numba==0.29.0
numexpr==2.4.3
numpy==1.11.2
pandas==0.19.1
protobuf==3.0.0
py==1.4.31
pylint==1.6.4
pypandoc==1.3.3
pytest==3.0.4
python-dateutil==2.6.0
pytz==2016.7
PyYAML==3.12
scikit-learn==0.18.1
scipy==0.17.0
six==1.10.0
sklearn-deap==0.1.7
tensorflow==0.11.0rc2
wrapt==1.10.8
yamlordereddictloader==0.1.1
